Refactor tree sitter parser code.
Only keep the API in parser.py.
Move the accessory functions into `parser_core.py`.
